+++ C/hungry_birds_threads.c
@@ -45,11 +45,11 @@ void *baby_bird(void *arg)
 		worms--;
 		if (worms) {
 			printf(" - baby bird %d eats (dish: %d worms)\n",
-				(int)arg, worms);
+				(int)(intptr_t)arg, worms);
 			pthread_mutex_unlock(&lock);
 		} else {
 			printf(" - baby bird %d eats (dish: %d worms)"
-				" and screams\n\n", (int)arg, worms);
+				" and screams\n\n", (int)(intptr_t)arg, worms);
 			pthread_mutex_unlock(&lock);
 			sem_post(&empty);
 		}
@@ -71,9 +71,9 @@ int main(int argc, char *argv[])
 
 	pthread_attr_init(&attr);
 	pthread_attr_setscope(&attr, PTHREAD_SCOPE_SYSTEM);
-	pthread_create(&parent_tid, &attr, parent_bird, (void *)0);
+	pthread_create(&parent_tid, &attr, parent_bird, (void *)(intptr_t)0);
 	for (i = 0; i < NUM_BIRDS; i++)
-		pthread_create(&baby_tid[i], &attr, baby_bird, (void *)i);
+		pthread_create(&baby_tid[i], &attr, baby_bird, (void *)(intptr_t)i);
 
 	pthread_join(parent_tid, NULL);
 	for (i = 0; i < NUM_BIRDS; i++)
